requirements.txt
boto3
click
pydantic
pyyaml
rich
mypy
pytest
